+++ Changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project follows to [Ragnarök Versioning Convention](https://shor.cz/ragnarok_versioning_convention).
 
+## [UNRELEASED] Red Core Version 0.6 Changelog
+
+### Red Core MC
+
+#### Added
+
+- `AABBUtil` a utility class for handling axis aligned bounding boxes
+- `AABBUtil#rotateAABB` a utility method for rotating a north facing AABB towards any `EnumFacing` allowing you to make only one AABB that works for every direction
+
 ## Red Core Version 0.5.1 Changelog - 2023-11-20
 
 ### Red Core MC